src/skeleton/views.py
import json
import rest_framework
from rest_framework.generics import RetrieveAPIView
from rest_framework.permissions import IsAuthenticated
from rest_framework.decorators import (
    authentication_classes,
    permission_classes,
    api_view,
)
from rest_framework.settings import api_settings
from django.http import HttpResponseNotAllowed
