requirements.txt
django==3.2.5
djangorestframework==3.12.4
django-extensions==3.1.3
django-cors-headers==3.7.0
django-simple-history==3.0.0
psycopg2-binary==2.9.1
django-filter==2.4.0
markdown==3.3.4
factory-boy==3.2.0
django-test-migrations==1.1.0
pydantic==1.8.2
werkzeug==2.0.1
flake8==3.9.2
python-gnupg==0.4.7
requests==2.25.1
PyJWT==1.7.1
# static files helper
whitenoise
# Unit testing packages
python-dotenv==0.18.0
pytest==6.2.4
pytest-django==4.4.0
pytest-factoryboy==2.1.0
pytest-pythonpath==0.7.3
# Formatters
black==21.5b1
# Typechecking packages
mypy==0.910
types-PyJWT==1.7.1
django-stubs==1.8.0
djangorestframework-stubs==1.4.0
django-filter-stubs==0.1.1
types-markdown
types-six
types-requests
types-Werkzeug
types-cryptography
types-toml